backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
numpy>=1.24.0
pandas>=2.0.0 
scikit-learn>=1.3.0 
scipy>=1.11.0
networkx
pdfplumber
reportlab

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10 
Pillow>=10.0.0 
opencv-python-headless>=4.8.0 
spacy>=3.7.0
python-Levenshtein>=0.20.0